lib/features/agents/presentation/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai>=0.5.0
pandas==2.2.0
numpy==1.26.3
TA-Lib==0.4.28
python-dotenv==1.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
apscheduler==3.10.4
pydantic==2.6.0
pydantic-settings==2.1.0
websockets==12.0
httpx==0.26.0